move pages to common, break up welcome page
